[Feat] Kill 업적 기능 구현
AchievementManager를 통해 킬 관련된 업적 기능 구현
