+++ Assets/LHJ/LHJ_SCENE/TestScene2.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0a073b10d581a9646bd1deb4489a5b91
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
