+++ Assets/LHJ/LHJ_Scripts/Test/TestKill.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 114345b58c6482f4ba6c9e782febceaf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
