+++ Assets/Prefabs/AchievementManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7d3b3f755607540408b820580cc82f4b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
